Allow system=java to prevent downloading a JDK
Turns out coursier already interprets the `system` value
to reference the system installed java runtime.

So, for example if you execute:

`coursier java-home --jvm system`

It will try to find your current JAVA_HOME.

We now take advantage of this and if users specify the
`java_version=system` we will make sure they have an
existing `java` command already available.

Also, we will try to download the small jvm-less coursier
script, instead of downloading it with a jre which will save
a lot of bandwidth.

The test for this sets a custom COURSIER_JVM_CACHE variable
and makes sure it we wont download a jvm inside of it.